requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pytube>=15.0.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth>=2.23.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.4
accelerate>=0.21.0
tqdm>=4.66.0
pyyaml>=6.0
watchdog>=3.0.0
loguru>=0.7.0
python-slugify>=8.0.1